Add files via upload
App has two classes:
*************************************Better.class*************************************
Takes two Strings and a lambda  that says whether the first or the two is “better”.
The method will return that better String; i.e.,
if the function given by the lambda returns true, the betterString method will return the first String,
otherwise betterString should return the second String.

*************************************CheckString*************************************
Checks if the string is empty or null or not. If empty or null , returns false
If the string is neither empty nor null, 
then checks using Lambda Expression whether the string contains only alphabets or not.